requirements.txt
-f /usr/share/pip-wheels
asgiref==3.3.4
backcall==0.2.0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
decorator==5.0.9
Django==3.2.4
django-category==2.0.1
django-crispy-forms==1.12.0
django-object-tools==2.0.0
django-redis-cache==2.1.0
django-rq==2.1.0
django-summernote==0.8.11.6
django-translation-manager==1.1.1
djangorestframework==3.12.4
idna==2.10
importlib-metadata==4.5.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
Markdown==3.3.4
mysqlclient==2.0.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
polib==1.1.1
prompt-toolkit==3.0.19
ptyprocess==0.7.0
Pygments==2.9.0
python-dotenv==0.18.0
pytz==2021.1
redis==3.5.3
requests==2.25.1
rq==1.9.0
six==1.16.0
sqlparse==0.4.1
traitlets==4.3.3
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
zipp==3.4.1
